Bug fix for where filter in projects page
Liquid where filter checking on null would not work with projects
collections when attribut was not present at all in Jekyll 2.3.1. Added
attribute (subpage) with default entry to false for all pages to avoid
the problem for now.
Also removed empty line in _config.yml to even show the collection.